backend/requirements.txt
pillow>=10.0.0                # For PNG handling
fastapi>=0.104.1             # Web server
uvicorn>=0.24.0             # Run the server
python-multipart>=0.0.6      # File uploads
pyinstaller>=6.0.0           # Create executable
pydantic>=2.4.2             # Data validation
send2Trash>=1.8.3           # Native delete functionality
psutil>=5.9.0               # Process management for KoboldCPP
sse-starlette>=1.6.5        # Server-sent events for streaming progress
requests>=2.31.0            # HTTP requests
aiofiles>=0.8.0             # Asynchronous file operations
SQLAlchemy>=1.4               # ORM for database interaction (SQLite)
pytest-asyncio>=0.21.0      # For asynchronous testing with pytest
pytest-cov>=2.12.1            # For test coverage
pytest-mock>=3.6.1            # For mocking in tests